Move tests from util to polyhedron.
